requirements.txt
meshtastic
requests
rich
python-dotenv
